+++ z3/src/z3_sort.cpp
@@ -61,12 +61,12 @@ SortVec Z3Sort::get_domain_sorts() const
 {
   if (is_function)
   {
-    int32_t s_arity = z_func.arity();
+    unsigned s_arity = z_func.arity();
     SortVec sorts;
     sorts.reserve(s_arity);
     Sort s;
 
-    for (size_t i = 0; i < s_arity; i++)
+    for (unsigned i = 0; i < s_arity; i++)
     {
       s.reset(new Z3Sort(z_func.domain(i), *ctx));
       sorts.push_back(s);
